Test change to optimize cast handling for movies
It seems that GetCast is fairly expensive, so add a flag to indicate if
cast members are desired or not.

This allows the JSON API to not request cast members (and the processing
involved) when not needed.

Also optimize the GetCast code to remove the looping test and replace
with a set lookup.  This should reduce O(n^2) loop.  This seems to make
marginal difference as the URL parsing is actually the expensive part.

Some clean up still todo, eg remove the old code (in the #elseif)