Speedup FileItem::IsSamePath
IsSamePath has been found to be expensive if PR 4301 is applied and
used.

Change IsSamePath to do a cheaper != null comparison first in a number
of && clauses and switch IsMusicDB and IsVideoDB to use the URIUtils
versions.  This stops building an expensive URL parser just to check the
protocol.

Note that a further optimization could be applied in cases where both
items have InfoTags, however, I'm not sure the extra optimization will
be needed.