Add HasAddon to AddonManager
Note that this isn't complete "test", but it does reduce some of the
overhead of testing if an addon is available in every frame.  Note
doesn't handle pvr addons.

The main case for this is on the confluence skin, it likes to test if
the global search addon is available on every frame.