Extend videodatabase GetTvShowInfo api
Add an optional FileItem *, this allows the retrieval of all the
properties from a tvshow, as some are only set when a fileitem is
available.

For the Hoff we probably should extend the videoinfotag to provide the
info without needing a fileitem, as it seems messy to not have all the
data in the info tag.

This allows the JSON-RPC GetTVShowDetails to return watchedepisodes, and
fixes this bug:
http://forum.xbmc.org/showthread.php?tid=191142