Fix CxMemFile so memory ownership is clear
The usage of CxMemFile is rather messy, and it's unclear who should own
the memory being passed into it.

Some code paths expect CxMemFile to handle memory, and some don't expect
it to free the memory.

Add explicit flag to indicate desired behaviour on constructor.

Add constructor with 0 params, and remove default params from other
constructor, I can't see a case where we should allow a pointer and a
default size of 0.

Update Alloc to honor buffer ownership.  If we realloc a buffer the
caller owns things could get messy.